+++ server/src/app.ts
@@ -3,11 +3,13 @@ import cors from "cors";
 import helmet from "helmet";
 import morgan from "morgan";
 
-import restaurantRoutes from "./routes/restaurant.routes";
-import qrRoutes from "./routes/qr.routes";
-import menuRoutes from "./routes/menu.routes";
-import orderRoutes from "./routes/order.routes";
-import categoryRoutes from "./routes/category.routes";
+import restaurantRoutes from "./routes/customer/restaurant.routes";
+import qrRoutes from "./routes/customer/qr.routes";
+import menuRoutes from "./routes/customer/menu.routes";
+import orderRoutes from "./routes/customer/order.routes";
+
+import categoryRoutes from "./routes/admin/category.routes";
+import menuItemRoutes from "./routes/admin/menuItem.routes";
 
 const app = express();
 
@@ -26,6 +28,8 @@ app.use("/api/restaurants", restaurantRoutes);
 app.use("/api/qr", qrRoutes);
 app.use("/api/restaurants", menuRoutes);
 app.use("/api/orders", orderRoutes);
+
 app.use("/api/admin/restaurants", categoryRoutes);
+app.use("/api/admin/restaurants", menuItemRoutes);
 
 export default app;